게시판 글 수정, 삭제 기능 구현 메인 페이지 제작

+++ src/main/webapp/view.jsp
@@ -144,8 +144,8 @@ a, a:hover {
 					<%
 						if(userID !=null && userID.equals(bbs.getUserID())){
 							%>
-								<a href="update.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">수정</a>
-								<a href="deleteAction.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">삭제</a>
+								<a onclick="return confirm('정말로 수정하시겠습니까?')" href="update.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">수정</a>
+								<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="deleteAction.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">삭제</a>
 							<%
 						}
 					%>
